test: Move taxonomyTermQuery to __utils__
When a test file imports another one their tests are also executed when
the file is called => So we move it to `__tests__/__utils__/`
