+++ __tests__/__utils__/index.ts
@@ -3,6 +3,7 @@ import * as R from 'ramda'
 export * from './assertions'
 export * from './handlers'
 export * from './services'
+export * from './query'
 
 export function getTypenameAndId(value: { __typename: string; id: number }) {
   return R.pick(['__typename', 'id'], value)
